# Conflicts:
#	doc/usage_command_line.md
#	file_lists/source_files_no_main
#	file_lists/test_files
#	src/stim.h
#	src/stim/circuit/circuit.cc
#	src/stim/cmd/command_diagram.cc
#	src/stim/cmd/command_diagram.h
#	src/stim/cmd/command_help.cc
#	src/stim/diagram/gate_data_svg.h
#	src/stim/diagram/timeline/timeline_svg_drawer.cc
#	src/stim/diagram/timeline/timeline_svg_drawer.h
#	src/stim/diagram/timeline/timeline_svg_drawer.test.cc
#	src/stim/main_namespaced.cc

+++ src/stim/diagram/timeline/timeline_svg_drawer.h
@@ -32,7 +32,7 @@ struct DiagramTimelineSvgDrawer {
     CircuitTimelineHelper resolver;
 
     size_t cur_moment = 0;
-    size_t moment_width = 1;
+    uint16_t moment_width = 1;
     size_t cur_moment_is_used = false;
     size_t tick_start_moment = 0;
     std::vector<bool> cur_moment_used_flags;
